+++ src/main/kotlin/fr/xpdustry/javelin/Javelin.kt
@@ -1,7 +1,6 @@
 package fr.xpdustry.javelin
 
 import arc.Core
-import arc.util.Strings
 import cloud.commandframework.annotations.AnnotationParser
 import cloud.commandframework.arguments.standard.StringArgument
 import cloud.commandframework.services.State
@@ -9,23 +8,26 @@ import com.google.inject.AbstractModule
 import com.google.inject.Guice
 import com.google.inject.Injector
 import com.google.inject.Provides
+import fr.xpdustry.distributor.Distributor
 import fr.xpdustry.distributor.command.ArcCommandManager
 import fr.xpdustry.distributor.command.ArcMeta
 import fr.xpdustry.distributor.command.sender.ArcCommandSender
 import fr.xpdustry.distributor.message.MessageIntent
 import fr.xpdustry.distributor.plugin.AbstractPlugin
-import fr.xpdustry.javelin.command.ServerCommand
+import fr.xpdustry.javelin.command.JavelinClientCommand
+import fr.xpdustry.javelin.command.JavelinServerCommand
 import fr.xpdustry.javelin.internal.JavelinClientConfig
 import fr.xpdustry.javelin.internal.JavelinServerConfig
 import fr.xpdustry.javelin.repository.ClientRepository
-import fr.xpdustry.javelin.service.JavelinWhisperService
-import fr.xpdustry.javelin.service.WhisperContext
-import fr.xpdustry.javelin.service.WhisperFormatter
-import fr.xpdustry.javelin.service.WhisperService
-import fr.xpdustry.javelin.util.format
+import fr.xpdustry.javelin.service.chat.GlobalChatContext
+import fr.xpdustry.javelin.service.chat.GlobalChatFormatter
+import fr.xpdustry.javelin.service.chat.GlobalChatService
+import fr.xpdustry.javelin.service.chat.JavelinGlobalChatService
+
+import fr.xpdustry.javelin.util.invoke
 import fr.xpdustry.javelin.util.formatter
-import fr.xpdustry.javelin.util.servicePipeline
 import fr.xpdustry.javelin.util.typeToken
+import mindustry.gen.Call
 import net.mindustry_ddns.store.FileStore
 import java.nio.charset.StandardCharsets
 import java.util.*
@@ -35,11 +37,11 @@ import javax.crypto.KeyGenerator
 class Javelin : AbstractPlugin() {
     companion object {
         @JvmStatic
-        var client: JavelinClient? = null
+        lateinit var client: JavelinClient
             private set
 
         @JvmStatic
-        var server: JavelinServer? = null
+        lateinit var server: JavelinServer
             private set
     }
 
@@ -52,51 +54,42 @@ class Javelin : AbstractPlugin() {
         serverStore = getStoredConfig("server-config", JavelinServerConfig::class.java)
         injector = Guice.createInjector(JavelinModule())
 
-        servicePipeline.registerServiceType(typeToken<WhisperService>(), WhisperService.local())
+        server = injector.getInstance(JavelinServer::class.java)
+        client = injector.getInstance(JavelinClient::class.java)
+
+        Distributor.getServicePipeline().registerServiceType(
+            typeToken<GlobalChatService>(),
+            JavelinGlobalChatService()
+        )
 
         if (serverStore.get().enabled) {
-            server = injector.getInstance(JavelinServer::class.java)
             Core.app.addListener(server)
         }
 
         if (clientStore.get().enabled) {
-            client = injector.getInstance(JavelinClient::class.java)
             Core.app.addListener(client)
-            servicePipeline.registerServiceImplementation(
-                typeToken<WhisperService>(),
-                injector.getInstance(JavelinWhisperService::class.java),
-                emptyList()
-            )
         }
     }
 
     override fun registerClientCommands(manager: ArcCommandManager) {
-        manager.command(manager.commandBuilder("whisper", "w")
-            .meta(ArcMeta.DESCRIPTION, "Whisper to somebody, I dunno...")
-            .argument(StringArgument.quoted("receiver"))
+        manager.command(manager.commandBuilder("global", "g")
+            .meta(ArcMeta.DESCRIPTION, "Send a message globally to all servers.")
+            .meta(ArcMeta.PARAMETERS, "[message...]")
             .argument(StringArgument.greedy("message"))
             .handler {
-                if (Strings.stripColors(it.sender.player.name()) == Strings.stripColors(it["receiver"])) {
-                    it.sender.sendMessage(it.sender.formatter.format(
-                        "You can't message yourself.", MessageIntent.ERROR)
-                    )
-                } else {
-                    val context = WhisperContext(it.sender.player.name(), it["receiver"], it["message"])
-                    val result = try {
-                        servicePipeline.pump(context).through(typeToken<WhisperService>()).result
-                    } catch (e: Exception) {
-                        State.REJECTED
-                    }
+                val context = GlobalChatContext(it.sender.player.name(), it["message"])
+                val result = try {
+                    Distributor.getServicePipeline().pump(context).through(typeToken<GlobalChatService>()).result
+                } catch (e: Exception) {
+                    State.REJECTED
+                }
 
-                    if (result == State.ACCEPTED) {
-                        it.sender.player.sendMessage(WhisperFormatter.instance.format(context))
-                    } else {
-                        it.sender.sendMessage(
-                            it.sender.formatter.format(
-                                "The player ${context.receiver} is not online.", MessageIntent.ERROR
-                            )
-                        )
-                    }
+                if (result == State.ACCEPTED) {
+                    Call.sendMessage(GlobalChatFormatter.instance.format(context))
+                } else {
+                    it.sender.sendMessage(it.sender.formatter.invoke(
+                        "Failed to broadcast the message. Please, report it to the server owner.", MessageIntent.ERROR
+                    ))
                 }
             }
         )
@@ -104,7 +97,8 @@ class Javelin : AbstractPlugin() {
 
     override fun registerServerCommands(manager: ArcCommandManager) {
         val annotations = AnnotationParser(manager, ArcCommandSender::class.java) { manager.createDefaultCommandMeta() }
-        annotations.parse(injector.getInstance(ServerCommand::class.java))
+        annotations.parse(injector.getInstance(JavelinServerCommand::class.java))
+        annotations.parse(injector.getInstance(JavelinClientCommand::class.java))
 
         manager.command(manager.commandBuilder("javelin").literal("server").literal("generate-secret")
             .meta(ArcMeta.DESCRIPTION, "Generates a secret key for your server, be aware that changing it revokes all you server tokens.")