+++ src/main/kotlin/fr/xpdustry/javelin/JavelinMessageHandler.kt
@@ -1,0 +1,5 @@
+package fr.xpdustry.javelin
+
+fun interface JavelinMessageHandler {
+    fun onMessageReceive(message: JavelinMessage, content: Any)
+}