+++ src/main/kotlin/fr/xpdustry/javelin/MessageHandler.kt
@@ -1,8 +1,0 @@
-package fr.xpdustry.javelin
-
-import fr.xpdustry.javelin.model.Endpoint
-
-interface MessageHandler {
-    val endpoint: Endpoint
-    fun onMessageReceive(message: JavelinMessage, content: Any)
-}