+++ src/main/kotlin/fr/xpdustry/javelin/service/WhisperContext.kt
@@ -1,7 +1,0 @@
-package fr.xpdustry.javelin.service
-
-data class WhisperContext(
-    val sender: String,
-    val receiver: String,
-    val message: String
-)