+++ src/main/kotlin/fr/xpdustry/javelin/service/chat/GlobalChatContext.kt
@@ -1,0 +1,6 @@
+package fr.xpdustry.javelin.service.chat
+
+data class GlobalChatContext(
+    val sender: String,
+    val message: String
+)