+++ src/main/kotlin/fr/xpdustry/javelin/service/chat/GlobalChatService.kt
@@ -1,0 +1,5 @@
+package fr.xpdustry.javelin.service.chat
+
+import cloud.commandframework.services.types.SideEffectService
+
+interface GlobalChatService : SideEffectService<GlobalChatContext>