+++ src/main/kotlin/fr/xpdustry/javelin/util/utils.kt
@@ -2,7 +2,6 @@
 
 package fr.xpdustry.javelin.util
 
-import cloud.commandframework.services.ServicePipeline
 import com.google.gson.Gson
 import fr.xpdustry.distributor.Distributor
 import fr.xpdustry.distributor.command.sender.ArcCommandSender
@@ -13,16 +12,16 @@ import org.java_websocket.drafts.Draft
 import org.java_websocket.drafts.Draft_6455
 import org.java_websocket.protocols.Protocol
 
-private val EMPTY_ARRAY = emptyArray<Any>()
-
-val servicePipeline: ServicePipeline
-    get() = Distributor.getServicePipeline()
-
 val ArcCommandSender.formatter: MessageFormatter
     get() = Distributor.getMessageFormatter(this)
 
-fun MessageFormatter.format(message: String, intent: MessageIntent = MessageIntent.INFO): String =
-    format(intent, message, EMPTY_ARRAY)
+operator fun MessageFormatter.invoke(
+    message: String,
+    intent: MessageIntent = MessageIntent.INFO,
+    vararg args: Any = emptyArray()
+): String {
+    return format(intent, message, *args)
+}
 
 inline fun <reified T> typeToken(): TypeToken<T> = object : TypeToken<T>() {}