+++ CHANGELOG.md
@@ -1,3 +1,11 @@
 - Cleanup + optimizations
 
 - Renamed `Server` to `Client`, and the former `ServerRepository` file `servers.json` to `clients.json`, just rename it...
+
+- Replaced the global whisper with a global chat, its more useful tbh...
+
+- `JavelinServer` and `JavelinClient` instances are guaranteed to be created on startup.
+
+- Added the ability to reconnect the client.
+
+- Improved documentation